upgrade lerna, creating private package for intergation tests, removing cycles from dependencies
