--- .jshintrc
@@ -1,4 +1,0 @@
-{
-	"esnext": true,
-	"asi": true
-}
