+++ packages/fela-dom/src/server/__tests__/renderToSheetList-test.js
@@ -1,5 +1,5 @@
-import renderToSheetList from '../renderToSheetList'
 import { createRenderer } from 'fela'
+import renderToSheetList from '../renderToSheetList'
 
 describe('Rendering to a sheet list', () => {
   it('should return a list of style sheets', () => {
