+++ packages/fela-utils/package.json
@@ -23,8 +23,5 @@
     "css-in-js-utils": "^3.0.0",
     "fast-loops": "^1.0.0",
     "string-hash": "^1.1.3"
-  },
-  "devDependencies": {
-    "fela": "^6.2.3"
   }
 }
