+++ packages/fela/package.json
@@ -36,8 +36,5 @@
     "fast-loops": "^1.0.0",
     "fela-utils": "^8.1.3",
     "isobject": "^3.0.1"
-  },
-  "devDependencies": {
-    "fela-tools": "^5.2.3"
   }
 }
