+++ packages/jest-fela-bindings/src/createSnapshotFactory.js
@@ -1,4 +1,3 @@
-import os from 'os'
 import { format } from 'prettier'
 import HTMLtoJSX from 'htmltojsx'
 
@@ -48,10 +47,11 @@ export default function createSnapshotFactory(
     )
 
     return (
-      formatCSS(renderToString(renderer)) +
-      os.EOL +
-      os.EOL +
-      formatHTML(div.innerHTML)
+      `${
+        formatCSS(renderToString(renderer))
+      }\n\n${
+        formatHTML(div.innerHTML)
+      }`
     )
   }
 }
